requirements.txt
requests~=2.26
peewee~=3.14.4
python-dotenv~=0.17.0
psycopg2-binary~=2.9.1
openpyxl~=3.0.7
beautifulsoup4~=4.10.0
tortoise-orm==0.17.6
asyncpg==0.24.0
pydantic==1.8.2
aiohttp~=3.8.1